\name{ew}
\alias{ew}
\docType{data}
\title{
Population Data for England and Wales
}
\description{
Population data for England and Wales combined between 1841 and 2007
}
\usage{data(ew)}
\format{
  The format is:
 Time-Series [1:167] from 1841 to 2007: 15845661 16021370 16201588 16399953 16606433 ...
}
\source{
The Human Mortality Database (http://www.mortality.org).
}
\references{
Human Mortality Database (2012). University of California, Berkeley (USA) and Max Planck
Institute for Demographic Research (Germany).
}
\examples{
data(ew)
plot(ew)
}
\keyword{datasets}
