\name{tsbugs-package}
\alias{tsbugs-package}
\alias{tsbugs}
\docType{package}
\title{
Create time series BUGS models.
}
\description{
The tsbugs package contains a collection of R functions that can be used to create time series BUGS models of various order. Included are function to create BUGS with non-constant variance such stochastic volatility models and random variance shift models.
}
\details{
\tabular{ll}{
Package: \tab tsbugs\cr
Type: \tab Package\cr
License: \tab GPL-2\cr
}
}
\author{
Guy J. Abel

Maintainer: Guy J. Abel <g.j.abel@gmail.com>
}
\references{
Blog posts with some additional details of the implementation of functions in the package can be found at:
\url{http://gjabel.wordpress.com/category/r/tsbugs/} 
}


